UPLOAD PAGE GENERATES CSV PREVIEW.
I added code that takes the CSV data and creates a table layout. We still need to let the user select the rows. The first row will become the header row and the rest will be come the data. 

Once we verify the CSV we can send the data back to the user. The user will be able to pair up the database columns to the columns in their CSV. 

Finally, the user will hit submit, and all the data will be imported into the data base. 

When adding data from a CSV, think of the process as a handshake. 
```
User => Server (upload/link file)
Server => user (acknowledge received data)
User => Server (verify rows)
Server => User (how to sort for database?)
User => Server (answer)
Server => Database (modify)
```