TESTS BACK TO NORMAL
I discovered some flaws in the GitHub actions. If the testing failed, the commit would still go through, but there would not be a discord notification. 

To fix it, I had to specifically state that I always wanted the discord notification job to run despite the outcome of the previous tests. 

As an added bonus, I figured out a way to get the status from other jobs. Since there are two jobs, 1. laravel-tests, and 2. notifications. The notifications job (aka Discord) is aware if the laravel-tests passed or failed. 

TL;DR: All tests work and discord displays if tests pass or fail.