Product model and money casting.
I create a model for products. Products will have the following data. Let me know if there should be anything else.

```
- id
- slug
- name
- cover_image
- quantity
- available
- currency
- price
- description
- scout comment
```

I plan to implement the following relation ships.
```
- sizes
- colors
- categories
- images
- merchants
- details / materials (may be 2 tables)
- reports
```

The most work was put into the model casting system. When the model implements a cast, the data from the database will automatically be converted to a custom PHP class. 

Without this the price on the model would just give me a number with no context. But now I can access all the methods and properties I may need for money. Such as easy conversion and displaying the symbol.

This screen shot shows how easy it is to convert money to another currency directly after pulling it out of the database.

https://imgur.com/tp8FWfX